LayoutTests/java/lc3/ConvertJSObject/ToLong-001-expected.txt

LiveConnect 3.0 JavaScript to Java Data Type Conversion Preferred argument conversion:  JavaScript Object to Long

FAIL TEST_CLASS.ambiguous( new String() ) +'' should be LONG. Was BYTE.
FAIL TEST_CLASS.ambiguous( new Boolean() ) +'' should be LONG. Was BYTE.
FAIL TEST_CLASS.ambiguous( new Number() ) +'' should be LONG. Was BYTE.
FAIL TEST_CLASS.ambiguous( new Date(0) ) +'' should be LONG. Was BYTE.
FAIL TEST_CLASS.ambiguous( new MyObject(999) ) +'' should be LONG. Was BYTE.
FAIL TEST_CLASS.ambiguous( MyFunction ) +'' should be LONG. Was BYTE.
PASS successfullyParsed is true

TEST COMPLETE
 
